fix(ci): resolve remaining compile/lint/typos failures
- gateway/telemetry.rs: use the SDK Tracer (stored from the OTLP provider)
  for the tracing-opentelemetry layer instead of the global BoxedTracer,
  which cannot satisfy PreSampledTracer in otel 0.21; construct
  SamplingResult { decision, .. } instead of the removed
  SamplingResult::RecordAndSample/Drop variants
- webhooks/dispatcher.rs: yield the failure message from the retry loop
  instead of a deferred mut Option<String> (fixes unused_mut warning)
- api/metrics.rs: Gauge::inc_by takes f64 (count as f64)
- tracing/db_tracing.rs: clone the span before Span::entered()
- tracing/kafka_propagator.rs: drop unnecessary mut
- time_series: rename Compressable/compressable -> Compressible/compressible
- .typos.toml: allow RTO and RECAL domain tokens

+++ src/api/metrics.rs
@@ -719,7 +719,7 @@ pub fn inc_job_scheduler_enqueued() {
 }
 
 pub fn inc_job_scheduler_claimed(count: u64) {
-    JOB_SCHEDULER_CLAIMED.inc_by(count);
+    JOB_SCHEDULER_CLAIMED.inc_by(count as f64);
 }
 
 pub fn inc_job_scheduler_heartbeat() {
